\name{textir-package}
\alias{textir}
\docType{package}
\title{Inverse Regression for Text}
\description{Tools for analysis of sentiment in text}
\details{
Check out Taddy (2011) and the help files.
}
\author{
Matt Taddy \email{taddy@chicagobooth.edu}
}
 \references{
\url{http://faculty.chicagobooth.edu/matt.taddy/index.html}

Taddy (2011), \emph{Inverse Regression for Analysis of Sentiment in Text}.
\url{http://arxiv.org/abs/1012.2098}

Taddy (2011), \emph{Estimation and Selection for Topic Models}.
\url{http://arxiv.org/abs/1109.4518}

}
\seealso{pls, mnlm}

