\name{tpx}
\alias{tpxfit}
\alias{tpxBF}
\alias{tpxweights}
\alias{tpxEM}
\alias{tpxML}
\alias{tpxQN}
\alias{tpxlpost}
\alias{tpxresids}
\alias{tpxThetaInit}
\alias{tpxOmegaInit}
\alias{tpxQ}
\alias{tpxHDet}
\alias{tpxToNEF}
\alias{tpxFromNEF}
\alias{tpxlogdet}
\title{
textir `tpx' functions.}
\description{ Undocumented internal utilities for topic model estimation. }
\references{ Taddy (2011), \emph{Estimation for Topic Models}.}
\author{
  Matt Taddy \email{taddy@chicagobooth.edu}
}
\seealso{topics}
