\name{tpx}
\alias{tpxinit}
\alias{tpxfit}
\alias{tpxSelect}
\alias{tpxweights}
\alias{tpxEM}
\alias{tpxML}
\alias{tpxQN}
\alias{tpxlpost}
\alias{tpxresids}
\alias{tpxThetaStart}
\alias{tpxOmegaStart}
\alias{tpxQ}
\alias{tpxMixQ}
\alias{tpxHnegDet}
\alias{tpxToNEF}
\alias{tpxFromNEF}
\alias{tpxlogdet}
\title{
textir `tpx' functions.}
\description{ Undocumented internal utilities for topic model estimation. }
\references{ Taddy (2011), \emph{Estimation for Topic Models}.}
\author{
  Matt Taddy \email{taddy@chicagobooth.edu}
}
\seealso{topics}
